+++ lib/image_io.dart
@@ -1,0 +1,4 @@
+export 'image.dart';
+export 'src/io/format_async.dart';
+export 'src/io/transform/copy_crop_async.dart';
+export 'src/io/transform/copy_resize_async.dart';