+++ test/transform/copy_crop_test.dart
@@ -16,4 +16,16 @@ void copyCropTest() {
       ..createSync(recursive: true)
       ..writeAsBytesSync(encodePng(i0_1));
   });
+
+  test('copyCropCircle', () {
+    final bytes = File('test/_data/png/buck_24.png').readAsBytesSync();
+    final i0 = PngDecoder().decodeImage(bytes)!.convert(numChannels: 4);
+    final i0_1 = copyCropCircle(i0);
+    expect(i0_1.width, equals(186));
+    expect(i0_1.height, equals(186));
+    expect(i0_1.format, equals(i0.format));
+    File('$testOutputPath/transform/copyCropCircle.png')
+      ..createSync(recursive: true)
+      ..writeAsBytesSync(encodePng(i0_1));
+  });
 }